Changes to make Rainloop compatible with NC14, as well as bringing it slightly more into the present by using the App Framework (to the extent possible).
