+++ rainloop/app/rainloop/v/1.14.0/app/libraries/RainLoop/Utils.php
@@ -540,7 +540,7 @@ class Utils
 	 */
 	public static function IsOwnCloudLoggedIn()
 	{
-		return self::IsOwnCloud() && \class_exists('OCP\User') && \OCP\User::isLoggedIn();
+		return self::IsOwnCloud() && \OC::$server->getUserSession()->isLoggedIn();
 	}
 
 	/**
@@ -562,10 +562,10 @@ class Utils
 		$sAppPath = '';
 		if (\RainLoop\Utils::IsOwnCloud())
 		{
-			if (\class_exists('OC_App'))
-			{
-				$sAppPath = \rtrim(\trim(\OC_App::getAppWebPath('rainloop')), '\\/').'/app/';
-			}
+			$sAppPath = \OC::$server->getURLGenerator()->linkToRoute('rainloop.page.appGet');
+			// TODO: Fix this ugly hack. Is there a "modern" way to return
+			// a URL that doesn't include index.php, which breaks things?
+			$sAppPath = preg_replace('index\.php\/', '', $sAppPath);
 
 			if (empty($sAppPath))
 			{
