+++ rainloop/appinfo/info.xml
@@ -21,4 +21,10 @@
   <screenshot>https://raw.githubusercontent.com/pierre-alain-b/rainloop-nextcloud/master/screenshots/2016.10.20-screenshot.jpg</screenshot>
 	<website>https://github.com/pierre-alain-b/rainloop-nextcloud</website>
 	<bugs>https://github.com/pierre-alain-b/rainloop-nextcloud/issues</bugs>
+	<screenshot>https://raw.githubusercontent.com/pierre-alain-b/rainloop-nextcloud/master/screenshots/2016.10.20-screenshot.jpg</screenshot>
+	<dependencies>
+		<php min-version="5.4" />
+		<owncloud min-version="6" max-version="9.2" />
+		<nextcloud min-version="10" max-version="14" />
+	</dependencies>
 </info>
